// src/three-types.d.ts
export interface IntersectionEvent<TSourceEvent> extends Intersection {
  /** The event source (the object which registered the handler) */
  eventObject: THREE.Object3D;
  /** An array of intersections */
  intersections: Intersection[];
  /** vec3.set(pointer.x, pointer.y, 0).unproject(camera) */
  unprojectedPoint: THREE.Vector3;
  /** Normalized event coordinates */
  pointer: THREE.Vector2;
  /** Delta between first click and this event */
  delta: number;
  /** The ray that pierced it */
  ray: THREE.Ray;
  /** The camera that was used by the raycaster */
  camera: Camera;
  /** stopPropagation will stop underlying handlers from firing */
  stopPropagation: () => void;
  /** The original host event */
  nativeEvent: TSourceEvent;
  /** If the event was stopped by calling stopPropagation */
  stopped: boolean;
}

export declare type Camera = THREE.OrthographicCamera | THREE.PerspectiveCamera;
export declare type ThreeEvent<TEvent> = IntersectionEvent<TEvent> &
  Properties<TEvent>;
export declare type DomEvent = PointerEvent | MouseEvent | WheelEvent;
export declare type EventHandlers = {
  onClick?: (event: ThreeEvent<MouseEvent>) => void;
  onContextMenu?: (event: ThreeEvent<MouseEvent>) => void;
  onDoubleClick?: (event: ThreeEvent<MouseEvent>) => void;
  onPointerUp?: (event: ThreeEvent<PointerEvent>) => void;
  onPointerDown?: (event: ThreeEvent<PointerEvent>) => void;
  onPointerOver?: (event: ThreeEvent<PointerEvent>) => void;
  onPointerOut?: (event: ThreeEvent<PointerEvent>) => void;
  onPointerEnter?: (event: ThreeEvent<PointerEvent>) => void;
  onPointerLeave?: (event: ThreeEvent<PointerEvent>) => void;
  onPointerMove?: (event: ThreeEvent<PointerEvent>) => void;
  onPointerMissed?: (event: MouseEvent) => void;
  onPointerCancel?: (event: ThreeEvent<PointerEvent>) => void;
  onWheel?: (event: ThreeEvent<WheelEvent>) => void;
};

export declare type Properties<T> = Pick<
  T,
  {
    [K in keyof T]: T[K] extends (_: any) => any ? never : K;
  }[keyof T]
>;
export declare type NonFunctionKeys<T> = {
  [K in keyof T]-?: T[K] extends Function ? never : K;
}[keyof T];
export declare type Overwrite<T, O> = Omit<T, NonFunctionKeys<O>> & O;
/**
 * If **T** contains a constructor, @see ConstructorParameters must be used, otherwise **T**.
 */
declare type Args<T> = T extends new (...args: any) => any
  ? ConstructorParameters<T>
  : T;
export declare type Euler = THREE.Euler | Parameters<THREE.Euler['set']>;
export declare type Matrix4 =
  | THREE.Matrix4
  | Parameters<THREE.Matrix4['set']>
  | Readonly<THREE.Matrix4['set']>;
/**
 * Turn an implementation of THREE.Vector in to the type that an r3f component would accept as a prop.
 */
declare type VectorLike<VectorClass extends THREE.Vector> =
  | VectorClass
  | Parameters<VectorClass['set']>
  | Readonly<Parameters<VectorClass['set']>>
  | Parameters<VectorClass['setScalar']>[0];
export declare type Vector2 = VectorLike<THREE.Vector2>;
export declare type Vector3 = VectorLike<THREE.Vector3>;
export declare type Vector4 = VectorLike<THREE.Vector4>;
export declare type Color =
  | ConstructorParameters<typeof THREE.Color>
  | THREE.Color
  | number
  | string;
export declare type ColorArray =
  | typeof THREE.Color
  | [color: THREE.ColorRepresentation];
export declare type Layers = THREE.Layers | Parameters<THREE.Layers['set']>[0];
export declare type Quaternion =
  | THREE.Quaternion
  | Parameters<THREE.Quaternion['set']>;
// ...
export declare type Node<T, P> = ExtendedColors<
  Overwrite<Partial<T>, NodeProps<T, P>>
>;
export declare type Object3DNode<T, P> = Overwrite<
  Node<T, P>,
  {
    position?: Vector3;
    up?: Vector3;
    scale?: Vector3;
    rotation?: Euler;
    matrix?: Matrix4;
    quaternion?: Quaternion;
    layers?: Layers;
    dispose?: (() => void) | null;
  }
> &
  EventHandlers;
export declare type BufferGeometryNode<
  T extends THREE.BufferGeometry,
  P
> = Node<T, P>;
export declare type MaterialNode<T extends THREE.Material, P> = Node<T, P>;
export declare type LightNode<T extends THREE.Light, P> = Object3DNode<T, P>;
export declare type Object3DProps = Object3DNode<
  THREE.Object3D,
  typeof THREE.Object3D
>;
export declare type AudioListenerProps = Object3DNode<
  THREE.AudioListener,
  typeof THREE.AudioListener
>;
export declare type PositionalAudioProps = Object3DNode<
  THREE.PositionalAudio,
  typeof THREE.PositionalAudio
>;
export declare type MeshProps = Object3DNode<THREE.Mesh, typeof THREE.Mesh>;
// ...
export declare type PerspectiveCameraProps = Object3DNode<
  THREE.PerspectiveCamera,
  typeof THREE.PerspectiveCamera
>;
export declare type OrthographicCameraProps = Object3DNode<
  THREE.OrthographicCamera,
  typeof THREE.OrthographicCamera
>;
// ...
export declare type MeshStandardMaterialProps = MaterialNode<
  THREE.MeshStandardMaterial,
  [THREE.MeshStandardMaterialParameters]
>;
export declare type MeshPhongMaterialProps = MaterialNode<
  THREE.MeshPhongMaterial,
  [THREE.MeshPhongMaterialParameters]
>;
export declare type MeshToonMaterialProps = MaterialNode<
  THREE.MeshToonMaterial,
  [THREE.MeshToonMaterialParameters]
>;
// ...
export declare type MeshBasicMaterialProps = MaterialNode<
  THREE.MeshBasicMaterial,
  [THREE.MeshBasicMaterialParameters]
>;
// ...
export declare type PointLightProps = LightNode<
  THREE.PointLight,
  typeof THREE.PointLight
>;
// ...
export declare type AmbientLightProps = LightNode<
  THREE.AmbientLight,
  typeof THREE.AmbientLight
>;
// ...
export declare type GridHelperProps = Object3DNode<
  THREE.GridHelper,
  typeof THREE.GridHelper
>;
// ...
export declare type AxesHelperProps = Object3DNode<
  THREE.AxesHelper,
  typeof THREE.AxesHelper
>;